+++ 3d/fmm.cxx
@@ -17,7 +17,7 @@ int main(int argc, char ** argv) {
   Bodies bodies(numBodies);                                     // Initialize bodies
   real_t average = 0;                                           // Average charge
   srand48(0);                                                   // Set seed for random number generator
-  for (int b=0; b<int(bodies.size()); b++) {                    // Loop over bodies
+  for (size_t b=0; b<bodies.size(); b++) {                      // Loop over bodies
     for (int d=0; d<3; d++) {                                   //  Loop over dimension
       bodies[b].X[d] = drand48() * 2 * M_PI - M_PI;             //   Initialize positions
     }                                                           //  End loop over dimension
@@ -27,7 +27,7 @@ int main(int argc, char ** argv) {
     for (int d=0; d<3; d++) bodies[b].F[d] = 0;                 //  Clear force
   }                                                             // End loop over bodies
   average /= bodies.size();                                     // Average charge
-  for (int b=0; b<int(bodies.size()); b++) {                    // Loop over bodies
+  for (size_t b=0; b<bodies.size(); b++) {                      // Loop over bodies
     bodies[b].q -= average;                                     // Charge neutral
   }                                                             // End loop over bodies
   stop("Initialize bodies");                                    // Stop timer
@@ -59,7 +59,7 @@ int main(int argc, char ** argv) {
   }                                                             // End loop over target samples
   bodies.resize(numTargets);                                    // Resize bodies
   Bodies bodies2 = bodies;                                      // Backup bodies
-  for (int b=0; b<int(bodies.size()); b++) {                    // Loop over bodies
+  for (size_t b=0; b<bodies.size(); b++) {                      // Loop over bodies
     bodies[b].p = 0;                                            //  Clear potential
     for (int d=0; d<3; d++) bodies[b].F[d] = 0;                 //  Clear force
   }                                                             // End loop over bodies
@@ -68,7 +68,7 @@ int main(int argc, char ** argv) {
 
   //! Verify result
   real_t pDif = 0, pNrm = 0, FDif = 0, FNrm = 0;
-  for (int b=0; b<int(bodies.size()); b++) {                    // Loop over bodies & bodies2
+  for (size_t b=0; b<bodies.size(); b++) {                      // Loop over bodies & bodies2
     pDif += (bodies[b].p - bodies2[b].p) * (bodies[b].p - bodies2[b].p);// Difference of potential
     pNrm += bodies2[b].p * bodies2[b].p;                        //  Value of potential
     FDif += (bodies[b].F[0] - bodies2[b].F[0]) * (bodies[b].F[0] - bodies2[b].F[0]) +// Difference of force
