+++ 3dp/traverse_lazy.h
@@ -60,7 +60,7 @@ namespace exafmm {
   //! Horizontal pass for periodic images
   void periodic(Cell * Ci0, Cell * Cj0, real_t cycle) {
     Cells pcells(27);                                           // Create cells
-    for (int c=0; c<int(pcells.size()); c++) {                  // Loop over periodic cells
+    for (size_t c=0; c<pcells.size(); c++) {                    // Loop over periodic cells
       pcells[c].M.resize(NTERM, 0.0);                           //  Allocate & initialize M coefs
       pcells[c].L.resize(NTERM, 0.0);                           //  Allocate & initialize L coefs
     }                                                           // End loop over periodic cells
